lexer: drop number token
The specification says:
> Symbols (any string not enclosed in double quotes) must only contain
> alphanumeric [a-z A-Z] [0-9] characters plus the following special
> characters: \.@=/-_$%@+!|&ˆ:

Hence, we should lex numbers as a symbol: numbers are a subset of the
symbol character set (specifically [0-9] x X . :) and are not enclosed
in quotes.  The parsing of numbers should be done later in the semantic
analysis stage; we don't have enough information to determine what kind
of number the token is without semantic context.

Signed-off-by: Rahul Sandhu <nvraxn@gmail.com>